chore(tests): mirror src/pinet tree, split parser unit tests (#138)
Reorganize the flat src/test/ into a tree mirroring src/pinet/ (issue
#130): constraint tests move under src/test/constraints/ and the VJP
test under src/test/solver/; top-level module tests (project,
dataclasses, equilibration, constraint_violation, runtime_typecheck,
benchmarks_smoke) stay at src/test/. Moves use `git mv` to preserve
history; tests import via the public `pinet` API so relocation needs no
import changes (pytest.ini testpaths/pythonpath unchanged).

Split: extract the self-contained ConstraintParser behavior tests
(identity passthrough, batch-size validation) out of the large
projection-equivalence files test_parser_and_affine_inequality.py and
test_parser_and_nonlinear.py into a dedicated
src/test/constraints/test_constraint_parser.py mirroring
constraints/constraint_parser.py.

No test logic changed; 605 tests pass (same count), ruff and
basedpyright --level error clean.

Relates to #130
