+++ src/test/constraints/test_clipped_sine.py
