+++ src/test/constraints/test_parser_and_affine_inequality.py
@@ -702,78 +702,3 @@ def test_affine_inequality_project_cannot_be_called_directly():
         ),
     ):
         ineq_constraint.project(ProjectionInstance(x=jnp.zeros((1, 2, 1))))
-
-
-def test_constraint_parser_no_ineq_no_box_returns_eq_as_is():
-    dim, n_eq = 3, 2
-    a_mat = jnp.arange(n_eq * dim, dtype=jnp.float64).reshape(1, n_eq, dim)
-    b = jnp.zeros((1, n_eq, 1))
-    eq = EqualityConstraint(a_mat=a_mat, b=b, method="pinv")
-
-    parser = ConstraintParser(eq_constraint=eq, ineq_constraint=None, box_constraint=None)
-    eq_out, box_out, _ = parser.parse(method="pinv")
-
-    # Still the same exact object (no lifting performed)
-    assert eq_out is eq, (
-        "Parser should return the original equality constraint when no lifting is needed."
-    )
-    assert box_out is None, (
-        "Parser should not synthesize a box constraint when none is provided."
-    )
-    assert eq_out is not None, (
-        "Parser should preserve the equality constraint when no lifting is needed."
-    )
-    assert eq_out.a_mat is a_mat, (
-        "Parser should preserve the original equality matrix when no lifting is needed."
-    )
-    assert eq_out.b is b, (
-        "Parser should preserve the original equality RHS when no lifting is needed."
-    )
-
-
-def test_constraint_parser_no_ineq_with_box_returns_inputs():
-    dim, n_eq = 4, 1
-    a_mat = jnp.ones((1, n_eq, dim))
-    b = jnp.zeros((1, n_eq, 1))
-    eq = EqualityConstraint(a_mat=a_mat, b=b, method="pinv")
-
-    mask = jnp.array([True, False, True, False])
-    n_box = int(mask.sum())
-    lb = jnp.array([[[-1.0], [0.0]]]).reshape(1, n_box, 1)
-    ub = jnp.array([[[1.0], [2.0]]]).reshape(1, n_box, 1)
-    box = BoxConstraint(BoxConstraintSpecification(lb=lb, ub=ub, mask=mask))
-
-    parser = ConstraintParser(eq_constraint=eq, ineq_constraint=None, box_constraint=box)
-    eq_out, box_out, _ = parser.parse(method="pinv")
-
-    # Still the same exact objects (no lifting performed)
-    assert eq_out is eq, (
-        "Parser should return the original equality constraint when only a box "
-        "constraint is provided."
-    )
-    assert box_out is box, (
-        "Parser should return the original box constraint when no lifting is needed."
-    )
-    assert isinstance(box_out, BoxConstraint), (
-        "Parser should preserve the box constraint when no lifting is needed."
-    )
-    assert box_out.mask is not None, (
-        "Preserved box constraint should still expose its mask."
-    )
-    assert box_out.lb is not None, (
-        "Preserved box constraint should still expose its lower bounds."
-    )
-    assert box_out.ub is not None, (
-        "Preserved box constraint should still expose its upper bounds."
-    )
-
-    # Sanity: mask/bounds unchanged
-    assert jnp.array_equal(box_out.mask, mask), (
-        "Parser should preserve the original box mask when no lifting is needed."
-    )
-    assert jnp.array_equal(box_out.lb, lb), (
-        "Parser should preserve the original box lower bounds when no lifting is needed."
-    )
-    assert jnp.array_equal(box_out.ub, ub), (
-        "Parser should preserve the original box upper bounds when no lifting is needed."
-    )
